requirements.txt
gspread>=6.0
google-auth>=2.20
requests>=2.28
python-dotenv>=1.0
